Create HappyModel EP82 VRX Target
This is a source of continuing confusion for users trying to flash a VRX backpack. It makes some sense to use the HappyModel EP target for an EP82, but in fact, those pins will result in the module not booting. This creates a new target for clarity when flashing these devices.